requirements.txt
python-dotenv==1.1.1
httpx==0.24.1
markdownify==1.1.0
pillow==11.3.0
pydantic==2.0.3
readabilipy==0.3.0
beautifulsoup4==4.13.4
fastmcp==2.11.2
